requirements.txt
api-client==1.3.1
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.6
django-filter==25.1
django-pytest==0.2.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
Markdown==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.5
pytest-django==4.10.0
requests==2.32.3
sqlparse==0.5.3
tenacity==9.0.0
urllib3==2.3.0
wheel==0.45.1
